fix: test class ModelService and refactor location test classes
