+++ app/src/test/java/com/example/stratego_app/ModelServiceTest.java
@@ -14,7 +14,7 @@ import com.example.stratego_app.model.Rank;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class modelServiceTestClass {
+public class ModelServiceTest {
 
     private ModelService modelService = new ModelService();
     private Board board = new Board();
@@ -23,6 +23,16 @@ public class modelServiceTestClass {
     public void setUp() {
         modelService = ModelService.getInstance();
         board = modelService.getGameBoard();
+    }
+
+    @Test
+    void getPieceTest() {
+        assertNull(modelService.getPieceAtPosition(1,1));
+    }
+
+    @Test
+    void getBoard() {
+        assertNotNull(modelService.getGameBoard());
     }
 
     @Test
@@ -40,9 +50,8 @@ public class modelServiceTestClass {
         boolean result = modelService.placePieceAtGameSetUp(5, 6, piece);
 
         assertTrue(result);
-        assertNotNull(modelService.getPieceAtPosition(5, 6));
         assertNotNull(board.getField(5,6));
-        assertEquals(piece, modelService.getPieceAtPosition(5,6));
+
     }
 
     @Test
